Separate mothership and component recon binaries
