+++ Makefile
@@ -27,24 +27,30 @@ build: build-linux build-darwin build-linux-arm
 
 .PHONY: build-linux
 build-linux:
-	CGO_ENABLED=0 GOOS=linux GOARCH=amd64 go build -o ./bin/reconciler-linux $(FLAGS) ./cmd
+	CGO_ENABLED=0 GOOS=linux GOARCH=amd64 go build -o ./bin/reconciler-linux $(FLAGS) ./cmd/reconciler
+	CGO_ENABLED=0 GOOS=linux GOARCH=amd64 go build -o ./bin/mothership-linux $(FLAGS) ./cmd/mothership
 
 .PHONY: build-linux-arm
 build-linux-arm:
-	CGO_ENABLED=0 GOOS=linux GOARCH=arm64 go build -o ./bin/reconciler-arm $(FLAGS) ./cmd
+	CGO_ENABLED=0 GOOS=linux GOARCH=arm64 go build -o ./bin/reconciler-arm $(FLAGS) ./cmd/reconciler
+	CGO_ENABLED=0 GOOS=linux GOARCH=arm64 go build -o ./bin/mothership-arm $(FLAGS) ./cmd/mothership
 
 .PHONY: build-darwin
 build-darwin:
-	CGO_ENABLED=0 GOOS=darwin GOARCH=amd64 go build -o ./bin/reconciler-darwin $(FLAGS) ./cmd
+	CGO_ENABLED=0 GOOS=darwin GOARCH=amd64 go build -o ./bin/reconciler-darwin $(FLAGS) ./cmd/reconciler
+	CGO_ENABLED=0 GOOS=darwin GOARCH=amd64 go build -o ./bin/mothership-darwin $(FLAGS) ./cmd/mothership
 
 .PHONY: docker-build
 docker-build:
-	docker build -t $(APP_NAME):latest .
+	docker build -t $(APP_NAME)/mothership:latest -f Dockerfile.mr .
+	docker build -t $(APP_NAME)/component:latest -f Dockerfile.cr .
 
 .PHONY: docker-push
 docker-push:
-	docker tag $(APP_NAME) $(IMG_NAME):$(TAG)
-	docker push $(IMG_NAME):$(TAG)
+	docker tag $(APP_NAME)/mothership $(IMG_NAME)/mothership:$(TAG)
+	docker tag $(APP_NAME)/component $(IMG_NAME)/component:$(TAG)
+	docker push $(IMG_NAME)/mothership:$(TAG)
+	docker push $(IMG_NAME)/component:$(TAG)
 
 .PHONY: bump-primage
 bump-primage:
