+++ cmd/mothership/config/get/bucket/cmd.go
