+++ cmd/mothership/config/get/cmd.go
