+++ cmd/mothership/config/get/key/cmd.go
