+++ cmd/mothership/mothership/start/test/component-reconcilers.json
