+++ cmd/mothership/mothership/start/test/requests/create_cluster_invalid_kubeconfig.json
